convert to BEM css convention
