+++ package.json
@@ -26,7 +26,9 @@
     "@typescript-eslint/parser": "^5.30.4",
     "eslint": "8.19.0",
     "eslint-config-next": "12.2.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.30.1",
+    "prettier": "^2.8.7",
     "typescript": "4.7.4"
   }
 }
